+++ src/components/render-page/window-post-message-communicator/rendering-message.ts
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: AGPL-3.0-only
  */
 import type { ScrollState } from '../../editor-page/synced-scroll/scroll-props'
-import type { RendererFrontmatterInfo } from '../../common/note-frontmatter/types'
+import type { RendererFrontmatterInfo } from '../../../redux/note-details/types/types'
 
 export enum CommunicationMessageType {
   SET_MARKDOWN_CONTENT = 'SET_MARKDOWN_CONTENT',
