+++ src/redux/note-details/reducer.ts
@@ -5,15 +5,15 @@
  */
 
 import type { Reducer } from 'redux'
-import type { PresentFrontmatterExtractionResult } from '../../components/common/note-frontmatter/types'
-import type { NoteFrontmatter } from '../../components/common/note-frontmatter/note-frontmatter'
-import { createNoteFrontmatterFromYaml } from '../../components/common/note-frontmatter/note-frontmatter'
-import type { NoteDetails, NoteDetailsActions } from './types'
+import { createNoteFrontmatterFromYaml } from './raw-note-frontmatter-parser/parser'
+import type { NoteDetailsActions } from './types'
 import { NoteDetailsActionType } from './types'
-import { extractFrontmatter } from '../../components/common/note-frontmatter/extract-frontmatter'
+import { extractor } from './frontmatter-extractor/extractor'
 import type { NoteDto } from '../../api/notes/types'
 import { initialState } from './initial-state'
 import { DateTime } from 'luxon'
+import type { NoteDetails, NoteFrontmatter } from './types/note-details'
+import type { PresentFrontmatterExtractionResult } from './frontmatter-extractor/types'
 
 export const NoteDetailsReducer: Reducer<NoteDetails, NoteDetailsActions> = (
   state: NoteDetails = initialState,
@@ -74,7 +74,7 @@ const buildStateFromTaskListUpdate = (
  * @return An updated {@link NoteDetails} redux state.
  */
 const buildStateFromMarkdownContentUpdate = (state: NoteDetails, markdownContent: string): NoteDetails => {
-  const frontmatterExtraction = extractFrontmatter(markdownContent)
+  const frontmatterExtraction = extractor(markdownContent)
   if (!frontmatterExtraction.isPresent) {
     return {
       ...state,
@@ -84,14 +84,15 @@ const buildStateFromMarkdownContentUpdate = (state: NoteDetails, markdownContent
       frontmatter: initialState.frontmatter,
       frontmatterRendererInfo: initialState.frontmatterRendererInfo
     }
+  } else {
+    return buildStateFromFrontmatterUpdate(
+      {
+        ...state,
+        markdownContent: markdownContent
+      },
+      frontmatterExtraction
+    )
   }
-  return buildStateFromFrontmatterUpdate(
-    {
-      ...state,
-      markdownContent: markdownContent
-    },
-    frontmatterExtraction
-  )
 }
 
 /**
